refactor(website): split page composition out of the recipe module
design.ts was doing two jobs: holding the VOCABULARY (13 class recipes and
the scale) and ASSEMBLING it into whole pieces of page (pageHeader, ctaPanel).
The second job moves to lib/utils/ui.ts, which is the location AGENTS.md names
for exactly this: extract a repeating class bundle into a lib/utils/ui.ts
helper returning an html fragment, not into a CSS class.

They stay out of components/ deliberately. That folder means one custom
element per file, registered with a tag, shipped and hydrated. Neither of
these is interactive; both render at SSR and disappear, so making them
elements would ship JavaScript to register a tag that does nothing, and a
display-only component is elided by the framework regardless.

design.ts is also not a utils/ candidate. lib/utils/ is for browser-safe pure
helpers; a table of class recipes is an app-wide VALUE, which AGENTS.md places
at lib/*.ts alongside links.ts.

35 tests pass. webjs check caught the half-finished import rewrite mid-move,
which is the check earning its place.
